Refactor application structure by moving components to organized directories. Update imports in App, AuthPage, and Dashboard to reflect new paths.
