+++ src/components/auth/index.ts
@@ -1,0 +1,2 @@
+export { default as SignInForm } from './SignInForm';
+export { default as SignUpForm } from './SignUpForm';
