+++ src/components/features/index.ts
@@ -1,0 +1,1 @@
+export { default as RestaurantDisplay } from './RestaurantDisplay';
