+++ src/components/profile/index.ts
@@ -1,0 +1,3 @@
+export { default as PersonalInformation } from './PersonalInformation';
+export { default as AllergensSection } from './AllergensSection';
+export { default as AccountSection } from './AccountSection';
